% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AcFunc.R
\name{AcFunc}
\alias{AcFunc}
\title{AcFunc}
\usage{
AcFunc(Cc, Rd, Vcmax, Kc, Ko, O2, coef)
}
\arguments{
\item{Cc}{CO2 concentration at the site of carboxylation}

\item{Rd}{Day respiration}

\item{Vcmax}{Michaelis-Menten VMax for carboxylation}

\item{Kc}{Michaelis-Menten parameter, binding coeff for carbon}

\item{Ko}{Michaelis-Menten parameter, binding coeff for oxygen}

\item{O2}{Oxygen concentration}

\item{coef}{Passed from CoefFunc}
}
\description{
Generates the C-limited data
}
