% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AjFunc.R
\name{AjFunc}
\alias{AjFunc}
\title{AjFunc}
\usage{
AjFunc(Cc, aG, aS, Rd, j, gammastar, coef)
}
\arguments{
\item{Cc}{CO2 concentration at the site of carboxylation}

\item{aG}{The proportion of glycerate carbon that exits photorespiration as glycine}

\item{aS}{The proportion of glycerate carbon that exits photorespiration as serine}

\item{Rd}{Day respiration}

\item{j}{Maximum rate of electron transport given current conditions}

\item{gammastar}{Cc compensation point}

\item{coef}{passed from CoefFunc}
}
\description{
Generates the J-limited data
}
