% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ApFunc.R
\name{ApFunc}
\alias{ApFunc}
\title{ApFunc}
\usage{
ApFunc(Cc, aG, aS, Rd, Vcmax, j, TPU, gm, gammastar)
}
\arguments{
\item{Cc}{CO2 concentration at the site of carboxylation}

\item{aG}{The proportion of glycerate carbon that exits photorespiration as glycine}

\item{aS}{The proportion of glycerate carbon that exits photorespiration as serine}

\item{Rd}{Day respiration}

\item{Vcmax}{Michaelis-Menten VMax for carboxylation}

\item{j}{Maximum rate of electron transport given current conditions}

\item{TPU}{Rate of triose phosphate usage}

\item{gm}{Mesophyll conductance to carbon}

\item{gammastar}{Cc compensation point}
}
\description{
Generates the P-limited data
}
