% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcGammaStar.R
\name{calcGammaStar}
\alias{calcGammaStar}
\title{Provide the details on your plants rubisco kinetics to calculate gamma star.}
\usage{
calcGammaStar(c, dHa, tleaf = 25, oxygen = 21)
}
\arguments{
\item{c}{Rubisco constant}

\item{dHa}{Rubisco constant}

\item{tleaf}{Leaf temperature in celsius. Default 25}

\item{oxygen}{Oxygen percent. Default 21.}
}
\description{
Provide the details on your plants rubisco kinetics to calculate gamma star.
}
