% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoefFunc.R
\name{CoefFunc}
\alias{CoefFunc}
\title{CoefFunc}
\usage{
CoefFunc(aG, gammastar, Cc)
}
\arguments{
\item{aG}{Proportion of glycerate carbon released from photorespiration as glycine}

\item{gammastar}{The Cc compensation point for rubisco}

\item{Cc}{Concentration of carbon dioxide at the site of carboxylation}
}
\description{
Generates a little data which makes life a lot easier
}
