% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitAci.R
\name{fitACi}
\alias{fitACi}
\title{fitAci}
\usage{
fitACi(
  data,
  gammastar = 3.52,
  O2 = 21,
  initialGuess = NA,
  forceValues = c(NA, NA, NA, NA, NA, NA, NA),
  bound_l = c(1, 1, 1, 0.001, 0.001, 0, 0),
  bound_h = c(1000, 1000, 1000, 30, 30, 1, 0.75),
  name_assimilation = "A",
  name_ci = c("pCi", "Ci"),
  pressure = 101,
  tleaf = 25
)
}
\arguments{
\item{data}{The A/Ci data to be used}

\item{gammastar}{The Cc compensation point. Default: Tobacco.}

\item{O2}{The oxygen concentration in parts per hundred.}

\item{initialGuess}{Initial guesses for fitting. Will generate ok guesses automatically if not provided}

\item{forceValues}{A vector of forced values in order: VcMax, J, TPU, gm, Rd, aG, aS}

\item{bound_l}{The minimum values to be considered for each parameter}

\item{bound_h}{the maximum value to be considered for each parameter}

\item{name_assimilation}{The name given to the assimilation column in "data"}

\item{name_ci}{Name given to the Ci column in "data"}

\item{pressure}{Atmospheric pressure in kPa}
}
\description{
Fit A/Ci curves using forced values, and bounds
}
