% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitComplete.R
\name{fitComplete}
\alias{fitComplete}
\title{fitComplete}
\usage{
fitComplete(
  data,
  name_assimilation = "A",
  name_ci = c("pCi", "Ci"),
  gammastar = 3.52,
  O2 = 21,
  pressure = 101,
  tleaf = 25,
  initialGuess = NA,
  forceValues = c(NA, NA, NA, NA, NA, NA, NA),
  bound_l = c(1, 1, 1, 0.001, 0.001, 0, 0),
  bound_h = c(1000, 1000, 1000, 30, 30, 1, 0.75),
  ignoreTPU = F,
  maxiter = 250,
  Kc = exp(35.9774 - (80.99/(0.008314 * (273.15 + tleaf)))),
  Ko = exp(12.3772 - (23.72/(0.008314 * (273.15 + tleaf))))
)
}
\arguments{
\item{data}{The A/Ci data to be used}

\item{name_assimilation}{The name given to the assimilation column in "data"}

\item{name_ci}{Name given to the Ci column in "data"}

\item{gammastar}{The Cc compensation point. Default: Tobacco.}

\item{O2}{The oxygen concentration in parts per hundred.}

\item{pressure}{Atmospheric pressure in kPa}

\item{tleaf}{The leaf temperature, in celsius}

\item{initialGuess}{Initial guesses for fitting. Will generate ok guesses automatically if not provided}

\item{forceValues}{A vector of forced values in order: VcMax, J, TPU, gm, Rd, aG, aS}

\item{bound_l}{The minimum values to be considered for each parameter}

\item{bound_h}{the maximum value to be considered for each parameter}

\item{ignoreTPU}{Whether to fit TPU or not. Leave false if you don't know what you're doing!}

\item{maxiter}{Maximum number of iterations for fitting. Can run very slowly if you have a large dataset and high maxiter. Default 250}

\item{Kc}{Michaelis-menten kinetic parameter for carboxylation}

\item{Ko}{Michaelis-menten kinetic parameter for oxygenation}
}
\description{
Fit A/Ci curves using forced values and dictated bounds. Returns a list with the fit, the graph, and the table.
}
