% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genGuess.R
\name{genGuess}
\alias{genGuess}
\title{genGuess}
\usage{
genGuess(A_Data)
}
\arguments{
\item{A_Data}{Assimilation data. We don't need Ci for this actually}
}
\description{
Generate initial guesses for VcMax, j, TPU
}
