% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsPresetsdocumentation.R
\docType{data}
\name{gsPresets}
\alias{gsPresets}
\title{Rubisco kinetic parameters of many species}
\format{
A tibble with 20 rows and 3 columns.
}
\usage{
data(gsPresets)
}
\description{
A dataset containing the kinetic parameters and their temperature sensitivities, used to calculated gamma*
}
\details{
\itemize{
    \item plant. The species this data pertains to.
    \item c. The temperature scaling constant of gamma*
    \item dHa. The enthalpy of the reaction in kJ / mol
}
}
\keyword{dataset}
