% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpFile.R
\name{interpFile}
\alias{interpFile}
\title{interpFile}
\usage{
interpFile(l)
}
\arguments{
\item{l}{Location of the file to be interpreted}
}
\description{
Attempts to read in the data file provided, if fails returns NULL so that the program won't break...
}
