% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nameParams.R
\name{namaParams}
\alias{namaParams}
\alias{nameParams}
\title{nameParams}
\usage{
nameParams(params)
}
\arguments{
\item{params}{The fitting parameters}
}
\description{
Names relevant parameters and returns a tibble
}
