% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstituteTable.R
\name{reconstituteTable}
\alias{reconstituteTable}
\title{Reconstitute data and pick limitations}
\usage{
reconstituteTable(
  data,
  fitParams,
  tleaf = 25,
  name_assimilation = "A",
  name_ci = c("pCi", "Ci"),
  pressure = 101,
  gammastar = 3.52,
  O2 = 21
)
}
\arguments{
\item{data}{The original A/Ci data}

\item{fitParams}{The returned params data from the fitting function}

\item{tleaf}{The leaf temperature}

\item{name_assimilation}{The name given to assimilation column in "data"}

\item{name_ci}{The name given to the internal CO2 concentration column in "data"}

\item{pressure}{The atmospheric pressure in kPa}
}
\description{
Take fit data and send it back through to build a table with limiting curves
}
