% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smartFit.R
\name{smartFit}
\alias{smartFit}
\title{smartFit}
\usage{
smartFit(
  data_assimilation,
  data_ci,
  ci_as_pressure = T,
  gm_min = 0.5,
  gm_max = 20,
  gm_samples = 10,
  gammastar = 3.52,
  O2 = 21,
  pressure = 101,
  tleaf = 25,
  forceValues = c(NA, NA, NA, NA, NA, NA, NA),
  bound_l = c(1, 1, 1, 0.001, 0.001, 0, 0),
  bound_h = c(1000, 1000, 1000, 30, 30, 1, 0.75),
  ignoreTPU = F,
  maxiter = 250,
  Kc = exp(35.9774 - (80.99/(0.008314 * (273.15 + tleaf)))),
  Ko = exp(12.3772 - (23.72/(0.008314 * (273.15 + tleaf))))
)
}
\arguments{
\item{data_assimilation}{A vector of assimilation values}

\item{data_ci}{A vector of the Ci values.}

\item{ci_as_pressure}{Whether Ci should be treated as pressure (T) or ppm (F).}

\item{gm_min}{Lowest gm value sampled.}

\item{gm_max}{Highest gm value sampled.}

\item{gm_samples}{How many gm values are sampled}

\item{gammastar}{The Cc compensation point. Default: Tobacco.}

\item{O2}{The oxygen concentration in parts per hundred.}

\item{pressure}{Atmospheric pressure in kPa}

\item{tleaf}{The leaf temperature, in celsius}

\item{forceValues}{A vector of forced values in order: VcMax, J, TPU, gm, Rd, aG, aS. Include NAs where not forced, if you modify this!}

\item{bound_l}{The minimum values to be considered for each parameter}

\item{bound_h}{the maximum value to be considered for each parameter}

\item{ignoreTPU}{Whether to fit TPU or not. Leave false if you don't know what you're doing!}

\item{maxiter}{How many iterations to allow for the fitting algorithm. If you have a lot of points, large values for maxiter will take a long time!}

\item{Kc}{Michaelis-menten kinetic parameter for carboxylation}

\item{Ko}{Michaelis-menten kinetic parameter for oxygenation}
}
\description{
Fit A/Ci curves iteratively, sampling several different gm values. Fitting is done in two steps for
each gm in the range. Fitting is done initially with forced gm to get the shape right and then refit 
with the previous fit as the starting condition to optimize gm. This should help the fitting program reach 
the most accurate fit with the least amount of finagling.
}
