% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_runsetter.R
\name{ss_runsetter}
\alias{ss_runsetter}
\title{A narrow function to read through an appended dataset and attempt to automatically split it up.}
\usage{
ss_runsetter(df, threshold = 2, aslist = FALSE, column = "Time")
}
\arguments{
\item{df}{The data table (singular) to analyze}

\item{threshold}{The difference in the chosen column that is to be considered a new measurement.}

\item{aslist}{When FALSE, returns a dataframe with an ID column that distinguishes separate measurements. When TRUE, returns the same but divided into a list, with one element per measurement.}

\item{column}{The chosen column to study for differences in threshold. Defaults to "Time," used by me to differentiate different measurements taken in the same file but separated by time.}
}
\description{
Use on appended data files where multiple measurements are made in one file. Will create an "ID" column for each individual measurement, as determined by threshold.
When the difference in the chosen column is greater than the provided threshold it will decide that it is a new measurement.
}
