% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ApFunc.R
\name{ApFunc}
\alias{ApFunc}
\title{ApFunc}
\usage{
ApFunc(Cc, aG, aS, Rd, TPU, gammastar, coef)
}
\arguments{
\item{Cc}{CO2 concentration at the site of carboxylation}

\item{aG}{The proportion of glycerate carbon that exits photorespiration as glycine}

\item{aS}{The proportion of glycerate carbon that exits photorespiration as serine}

\item{Rd}{Day respiration}

\item{TPU}{Rate of triose phosphate usage}

\item{gammastar}{Cc compensation point}

\item{coef}{passed from CoefFunc}
}
\description{
Generates the P-limited data
}
