% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genFun.R
\name{genFun}
\alias{genFun}
\title{Create the function from which to fit}
\usage{
genFun(
  forceValues = c(NA, NA, NA, NA, NA, NA, NA),
  gammastar = 3.52,
  O2 = 21,
  pCi,
  assimilationData
)
}
\arguments{
\item{forceValues}{A vector of forced values in order: VcMax, J, TPU, gm, Rd, aG, aS}

\item{gammastar}{The Cc compensation point. Default: Tobacco.}

\item{O2}{The oxygen concentration in parts per hundred.}

\item{pCi}{The internal pressure of CO2}

\item{assimilationData}{The measured assimilation data}
}
\description{
This function accepts some of the parameters required to fit an A/Ci curve.
It also accepts a vector of forced data to fit with and generates a function that remembers which parameters are fixed
}
