% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstituteGraph.R
\name{reconstituteGraph}
\alias{reconstituteGraph}
\title{Reconstitute and Graph fit data}
\usage{
reconstituteGraph(
  data,
  fitParams,
  tleaf,
  name_assimilation = "A",
  name_ci = c("pCi", "Ci"),
  pressure = 101,
  gammastar = 3.52
)
}
\arguments{
\item{data}{The original A/Ci data}

\item{fitParams}{The returned params data from the fitting function}

\item{tleaf}{The leaf temperature}

\item{name_assimilation}{The name given to assimilation column in "data"}

\item{name_ci}{The name given to the internal CO2 concentration column in "data"}

\item{pressure}{The atmospheric pressure in kPa}
}
\description{
Take fit data and send it back through to make a graph
}
