% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstituteTable.R
\name{reconstituteTable}
\alias{reconstituteTable}
\title{Reconstitute data and pick limitations}
\usage{
reconstituteTable(
  data,
  fitParams,
  name_assimilation = "A",
  name_ci = c("pCi", "Ci"),
  gammastar = 3.52,
  O2 = 21,
  pressure = 101,
  tleaf = 25,
  ignoreTPU = F,
  Kc = exp(35.9774 - (80.99/(0.008314 * (273.15 + tleaf)))),
  Ko = exp(12.3772 - (23.72/(0.008314 * (273.15 + tleaf))))
)
}
\arguments{
\item{data}{The original A/Ci data}

\item{fitParams}{The returned params data from the fitting function}

\item{name_assimilation}{The name given to assimilation column in "data"}

\item{name_ci}{The name given to the internal CO2 concentration column in "data"}

\item{gammastar}{The Cc compensation point. Default: Tobacco.}

\item{O2}{The oxygen concentration in parts per hundred.}

\item{pressure}{Atmospheric pressure in kPa}

\item{tleaf}{The leaf temperature, in celsius}

\item{ignoreTPU}{Whether to fit TPU or not. Leave false if you don't know what you're doing!}

\item{Kc}{Michaelis-menten kinetic parameter for carboxylation}

\item{Ko}{Michaelis-menten kinetic parameter for oxygenation}
}
\description{
Take fit data and send it back through to build a table with limiting curves
}
